core/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
streamlit==1.35.0
numpy==1.26.4
scipy==1.11.4
soundfile==0.12.1
requests==2.31.0
python-multipart==0.0.9
pydantic==2.7.1
PyYAML==6.0.1
pytest==8.2.0
